+++ user-api/src/main/java/org/ecommerce/userapi/provider/JwtProvider.java
@@ -1,4 +1,4 @@
-package org.ecommerce.userapi.security;
+package org.ecommerce.userapi.provider;
 
 import java.util.Date;
 import java.util.List;
@@ -10,7 +10,6 @@ import javax.crypto.SecretKey;
 
 import org.ecommerce.common.error.CustomException;
 import org.ecommerce.userapi.exception.UserErrorCode;
-import org.ecommerce.userapi.provider.RedisProvider;
 import org.springframework.http.HttpHeaders;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
