+++ user-api/src/main/java/org/ecommerce/userapi/security/JwtFilter.java
@@ -3,6 +3,7 @@ package org.ecommerce.userapi.security;
 import java.io.IOException;
 
 import org.ecommerce.userapi.exception.UserErrorCode;
+import org.ecommerce.userapi.provider.JwtProvider;
 import org.ecommerce.userapi.provider.RedisProvider;
 import org.ecommerce.userapi.security.custom.ResponseConfigurer;
 import org.springframework.security.authentication.ProviderManager;
